Adding tdbfile.sh
This is a bash script I wrote to solve a problem at work. There's a file contained in /var/lib/samba/printing that contains cached output of the lqp command for each printer queue in CUPS. If these files exceed a certain size, they can cause delayed and inconsistent printing in certain types of receipt printers. I wrote a script that can be used as a cronjob to look through the directory where these files are stored and deletes them if they get too big for their own good.